+++ src/Pyros3D/Utils/ModelLoaders/MultiModelLoader/AnimationLoader.cpp
@@ -38,7 +38,7 @@ namespace p3d {
             int32 channelsSize;
             bin->Read(&channelsSize, sizeof(int32));
 
-            for (int32 k=0;k<channelsSize;k++)
+            for (uint32 k=0;k<channelsSize;k++)
             {
                 // Channel
             	Channel ch;
@@ -48,14 +48,14 @@ namespace p3d {
                 bin->Read(&channelNameSize, sizeof(int32));
                 ch.NodeName.resize(channelNameSize);
                 bin->Read(&ch.NodeName[0], sizeof(char)*channelNameSize);
-
+                
                 // Positions
                 int32 positionSize;
                 bin->Read(&positionSize, sizeof(int32));
                 ch.positions.resize(positionSize);
-                for (int l=0;l<positionSize;l++)
+                for (uint32 l=0;l<positionSize;l++)
                 {
-                    bin->Read(&ch.positions[l].Time, sizeof(f64));
+                    bin->Read(&ch.positions[l].Time, sizeof(f32));
                     bin->Read(&ch.positions[l].Pos, sizeof(Vec3));
                 }
                 
@@ -63,9 +63,9 @@ namespace p3d {
                 int32 rotationSize;
                 bin->Read(&rotationSize, sizeof(int32));
                 ch.rotations.resize(rotationSize);
-                for (int l=0;l<rotationSize;l++)
+                for (uint32 l=0;l<rotationSize;l++)
                 {
-                    bin->Read(&ch.rotations[l].Time, sizeof(f64));
+                    bin->Read(&ch.rotations[l].Time, sizeof(f32));
                     bin->Read(&ch.rotations[l].Rot, sizeof(Quaternion));
                 }
 
@@ -73,19 +73,20 @@ namespace p3d {
                 int32 scalingSize;
                 bin->Read(&scalingSize, sizeof(int32));
                 ch.scales.resize(scalingSize);
-                for (int l=0;l<scalingSize;l++)
+                for (uint32 l=0;l<scalingSize;l++)
                 {
-                    bin->Read(&ch.scales[l].Time, sizeof(f64));
+                    bin->Read(&ch.scales[l].Time, sizeof(f32));
                     bin->Read(&ch.scales[l].Scale, sizeof(Vec3));
                 }
 
                 animation.Channels.push_back(ch);
+
             }
             // Duration
-            bin->Read(&animation.Duration, sizeof(uint32));
-
+            bin->Read(&animation.Duration, sizeof(f32));
+            
             // Ticks Per Second
-            bin->Read(&animation.TicksPerSecond, sizeof(f64));
+            bin->Read(&animation.TicksPerSecond, sizeof(f32));
 
             // Add Animation
             animations.push_back(animation);
